+++ app/gui/view/documentation/tailwind.config.js
@@ -1,0 +1,8 @@
+/** @type {import('tailwindcss').Config} */
+module.exports = {
+    content: ['src/**/*.rs'],
+    theme: {
+        extend: {},
+    },
+    plugins: [],
+}